+++ action/install/scripts/install.sh
@@ -6,6 +6,13 @@ set -e
 printf "release: ${INPUT_RELEASE}\n"
 printf "branch: ${INPUT_BRANCH}\n"
 
+if [[ "${INPUT_DOINSTALL}" == "false" ]]; then
+    printf "Skipping install, spliced should be installed by you.\n"
+    exit 0;
+else
+    printf "Moving forward with install ${INPUT_DOINSTALL}\n"
+fi
+
 python -m pip install --upgrade pip setuptools wheel
 
 # Case 1: no branch or release, install from pip