+++ spliced/schemas.py
@@ -93,6 +93,9 @@ result_properties = {
             "properties": {
                 "original": libs,
                 "spliced": libs,
+                # If a splice fails, we still provide libs here
+                "dep": libs,
+                "replace": libs,
             },
             "spliced": {"type": "array", "items": {"type": "string"}},
         },